fix setup.sh and dofiles.py

+++ setup.sh
@@ -1,19 +1,46 @@
 #!/usr/bin/env bash
-# must install 
-sudo package manager gvim zsh tmux git curl emacs rsync npm tig cmake ctags
+# set package manager
+PKMG="package_manager"
+
+## --------------------------------------------must install -------------------------------------------- ##
+# common must install 
+sudo $PKMG zsh tmux git curl emacs rsync nodejs npm tig cmake ctags xsel
+# npm setting
+sudo npm config set registry https://registry.npm.taobao.org --global
+sudo npm config set disturl https://npm.taobao.org/dist --global
 # install fzf
 git clone --depth 1 https://github.com/junegunn/fzf.git ~/.fzf
 ~/.fzf/install
 
 # install proxychains-ng 
-git clone https://github.com/rofl0r/proxychains-ng.git
-cd proxychains-ng
-# configure and install 
-./configure --prefix=/usr --sysconfdir=/etc
-make
-sudo make install
-sudo make install-config # installs /etc/proxychains.conf
-sudo sed -i "s/.*socks4[[:space:]]*127.0.0.1.*/socks5  127.0.0.1 1080/g" /etc/proxychains.conf
+if hash proxychains4 2> /dev/null; then
+  echo proxychains is installed!
+
+else
+  git clone https://github.com/rofl0r/proxychains-ng.git
+  cd proxychains-ng
+  # configure and install 
+  ./configure --prefix=/usr --sysconfdir=/etc
+  make
+  sudo make install
+  sudo make install-config # installs /etc/proxychains.conf
+  cd ..
+  sudo rm proxychains-ng -r
+  sudo sed -i "s/.*socks4[[:space:]]*127.0.0.1.*/socks5  127.0.0.1 1080/g" /etc/proxychains.conf
+fi
+
+# choose install
+ARCH="pacman -S"
+DEBIAN="apt install"
+if [[ "$PKMG" = "$ARCH"  ]]; then
+  sudo $PKMG gvim ctags the_silver_searcher yaourt shadowsocks-qt5 
+elif [[ "$PKMG"="$DEBIAN" ]]; then
+  sudo add-apt-repository ppa:hzwhuang/ss-qt5
+  sudo apt-get update
+  sudo $PKMG vim-gtk exuberant-ctags silversearcher-ag shadowsocks-qt5
+fi
+
+# --------------------------------------------common config -------------------------------------------- #
 
 # set up top font
 curl -L https://github.com/hbin/top-programming-fonts/raw/master/install.sh | bash
@@ -22,15 +49,14 @@ curl -L https://github.com/hbin/top-programming-fonts/raw/master/install.sh | ba
 cp ./config/Xmodmap ~/.Xmodmap
 sudo sh -c 'echo "xmodmap ~/.Xmodmap" >> /etc/rc.local'
 
+# --------------------------------------------vim -------------------------------------------- #
+
+
 #配置vim vim8 并且须支持python3 javascript
 git clone https://github.com/liuchengxu/space-vim.git ~/.space-vim
 cp ./space-vim/spacevim ~/.spacevim
 cp ./space-vim/private ~/.space-vim/ -r
-cp ./UltiSnips/ ~/.vim/
-
-# by manual
-# cd ~/.space-vim
-# make vim 
+cp ./space-vim/UltiSnips/ ~/.vim/ -r
 
 # config ycmd
 # sudo git clone https://github.com/Valloric/YouCompleteMe.git ~/.vim/plugged/YouCompleteMe
@@ -39,22 +65,32 @@ cp ./UltiSnips/ ~/.vim/
 # sudo python3 ./install.py --clang-completer
 
 # python vim  config and include following config
-sudo pip install -r ./config/requirements.txt
+sudo pip3 install -r $PWD/config/requirements.txt
 
 # web vim
 
 # html 
-sudo package manager tidy #查错
+sudo $PKMG tidy #查错
 sudo npm i -g js-beautify #js 和html格式化
 # javascript 
 sudo npm i -g eslint jslint tern #查错 补全
 # vue 
 sudo npm i -g eslint-plugin-vue #查错
 
+
+# manual install 
+
+# cd ~/.space-vim
+# make vim 
+
+# --------------------------------------------emacs -------------------------------------------- #
+
 # config emacs
 git clone https://github.com/syl20bnr/spacemacs ~/.emacs.d
-cp ./pittcat-spacemacs/spacemacs ~/.spacemacs
-cp ./pittcat-spacemacs/private ~/.emacs.d/ -r
+cp $PWD/pittcat-spacemacs/spacemacs ~/.spacemacs
+cp $PWD/pittcat-spacemacs/private ~/.emacs.d/ -r
+
+# --------------------------------------------zsh -------------------------------------------- #
 
 
 # config zsh
@@ -64,14 +100,20 @@ sh -c "$(curl -fsSL https://raw.githubusercontent.com/robbyrussell/oh-my-zsh/mas
 # pip install -r ./config/requirements.txt
 
 # install autojump
-sudo package manager autojump
-git clone git://github.com/joelthelion/autojump.git
-cd autojump
-./install.py
-cd ..
-sudo rm autojump -r
+if hash autojump 2> /dev/null; then
+  echo autojump is installed!
+else
+  sudo $PKMG autojump
+  git clone git://github.com/joelthelion/autojump.git
+  cd autojump
+  ./install.py
+  cd ..
+  sudo rm autojump -r
+fi
 
 cp ./zshrc ~/.zshrc
+
+# --------------------------------------------tmux -------------------------------------------- #
 
 
 # config tmux
@@ -83,8 +125,11 @@ make
 sudo make install 
 cd ..
 sudo rm tmux-mem-cpu-load -r
-sudo package manager xsel
-cp tmux.conf ~/.tmux.conf
+
+
+cp $PWD/tmux.conf ~/.tmux.conf
+
+# --------------------------------------------manual operation -------------------------------------------- #
 
 # manual operation
 
@@ -95,13 +140,5 @@ cp tmux.conf ~/.tmux.conf
 # git clone git://github.com/zsh-users/zsh-autosuggestions $ZSH_CUSTOM/plugins/zsh-autosuggestions
 # git clone https://github.com/zsh-users/zsh-syntax-highlighting.git ${ZSH_CUSTOM:-~/.oh-my-zsh/custom}/plugins/zsh-syntax-highlighting
 
- # must install 
-# silversearcher-ag 
-
-# desktop
-sudo package manager plank chromium
-
-# plank 
-cp ./config/autostart/plank.desktop ~/.config/autostart/
